use develop branches for graviton/graviphoton
This will need to get reverted as soon as stable tags have been created in the repos.
For now I'm just letting this build a develop based vagrant box so that we can proceed
with development. In the end this will need to be done i a fashio that allows us to build
both develop/master and tag based vagrant boxen.